+++ app/src/core/components/gallery/gallery_settings/thumbnail_size_slider/message.rs
@@ -1,4 +1,4 @@
 #[derive(Debug, Clone)]
 pub enum Message {
-    ValueChanged(u32),
+    ValueChanged(u16),
 }
